ewm: main_svc_try_5

+++ stat-service/dto-service/src/main/java/ru/practicum/ewm/ViewStats.java
@@ -12,5 +12,5 @@ import lombok.ToString;
 public class ViewStats {
     private String app;
     private String uri;
-    private Integer hits;
+    private Long hits;
 }
